Rewrite operator[] to avoid gcc-8 codegen bug
An upcoming change to SkVMGenerator triggers a gcc-8 x86-only bug. This
change prevents that, by limiting the conversions back and forth from
size_t to int (particularly here, where the original value is also
stored using a bitfield).

+++ src/sksl/SkSLVMGenerator.cpp
@@ -72,8 +72,8 @@ struct Value {
         skvm::Val& fVal;
     };
 
-    ValRef    operator[](int i)       { return fVals[i]; }
-    skvm::Val operator[](int i) const { return fVals[i]; }
+    ValRef    operator[](size_t i)       { return fVals[i]; }
+    skvm::Val operator[](size_t i) const { return fVals[i]; }
 
     SkSpan<skvm::Val> asSpan() { return fVals; }
